Convert HMR client to TS
